Implemented the drag and drop UI for Binding Policy
